Create Points_for_solving_problem
e
Points for the task.
question
The site contains a set of programming tasks.

Problems have 3 levels of difficulty: "easy", "medium", "hard".

For solving a problem of the "easy" level 2 points are awarded, for solving a problem of the "medium" level 4 points are awarded, for solving a problem of the "hard" level 8 points are awarded.

Write a function scoreIncrease that takes two values ​​current (the current score) and level (the difficulty level of the problem) as arguments and returns the new score after solving the problem. You must use a ternary operator in your solution.