Create Fruits cost using ternary operator
Fruit cost
question
A fruit merchant sells mangoes for $ 1 apiece. But if the buyer takes 3 or more mangoes, then the seller asks for $ 0.8 for one fruit. And if they buy 10 pieces or more, the seller is ready to sell them for only 60 cents apiece.

Write a function saleFruits that takes n (the number of fruits) as an argument and returns the purchase price. Round off the result to 2 decimal places. You must use a ternary operator in your solution.