Create Switch.Robot.js
Robot Coordinates
question
The initial coordinates of the Robot are known (x, y). The robot can move in four directions: N - North, S - South, W - West, E - East. When moving in a certain direction from the point (x, y), the corresponding coordinate of the Robot changes by 1:

when moving to the north, the Robot is at the point (x, y + 1),
when moving to the south, the Robot is at the point (x, y-1),
when moving to the west, the Robot is at the point (x-1, y),
when moving to the east, the Robot is at the point (x + 1 y).
robot

Write a function named robot that takes three arguments: the numbers x, y (the starting coordinates of the Robot) and a string direction containing the direction of movement ('N', 'S', 'E', 'W') and returns the new coordinates of the Robot in as an array x, y.